- add a base class for the dialog panels
